requirements.txt
django-braces
Django
djangorestframework
redis
drf-nested-routers
# django[bcrypt]
drf-extensions
pygraphviz
pyparsing==1.5.7
pydot
django-extensions
